pyproject.toml
[tool.poetry]
name = "grzemail"
version = "0.1.0"
description = "A console email client to rival mutt"
authors = ["Grzegorz Koperwas <admin@grzegorzkoperwas.site>"]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.10"
keyring = "^23.7.0"
SQLAlchemy = "^1.4.39"
aiosmtplib = "^1.1.6"
argparse = "^1.4.0"
configparser = "^5.2.0"
aioimaplib-hakiergrzonzo = {path = "../aioimaplib"}
textual = "^0.1.18"
socpi = "^0.2.1"
aiosqlite = "^0.17.0"
aiofiles = "^0.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
